+++ src/http-client/got.ts
@@ -1,5 +1,4 @@
-import got from 'got';
-import { HttpClientConfig } from './types.js';
+import got, { Headers, Options } from 'got';
 
 export const auth = (token: string, secret: string) =>
   got.extend({
@@ -32,7 +31,9 @@ export const logging = (logRequest = false, logResponse = false) =>
             const method = response.method;
             const url = response.url?.toString() ?? '';
             const body = JSON.stringify(response.body);
+            const headers = JSON.stringify(response.headers);
             console.log(`[Response] ${method ?? '?'} ${url}`);
+            console.log(`  headers: ${headers}`);
             console.log(`  body: ${body}`);
           }
 
@@ -45,4 +46,33 @@ export const logging = (logRequest = false, logResponse = false) =>
 export const pcoApi = () =>
   got.extend({
     prefixUrl: 'https://api.planningcenteronline.com',
+    /* eslint-disable */
+    handlers: [
+      (options: Options, next: (options: Options) => any) => {
+        return (async () => {
+          try {
+            const response = await next(options);
+            response.rateLimit = getRateLimit(response.headers as Headers);
+            return response;
+          } catch (error: any) {
+            const { response } = error;
+            if (response) {
+              error.rateLimit = getRateLimit(response.headers);
+            }
+
+            throw error;
+          }
+        })();
+      },
+    ],
+    /* eslint-enable */
   });
+
+function getRateLimit(headers: Headers) {
+  return {
+    count: +(headers['x-pco-api-request-rate-count'] as string),
+    limit: +(headers['x-pco-api-request-rate-limit'] as string),
+    period: +(headers['x-pco-api-request-rate-period'] as string),
+    retryAfter: +(headers['retry-after'] as string),
+  };
+}